braintrust/pyproject.toml
[project]
name = "btust"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "braintrust[cli]>=0.12.1",
    "modaic>=0.29.1",
    "modaic-client>=0.29.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[tool.uv.sources]
modaic-client = {path = "/Users/tytodd/Desktop/Modaic/code/core/modaic/src/modaic-client", editable = true}
